extract mothership-proxy concern
