+++ packages/unraid-api-plugin-connect/src/mothership-proxy/mothership.module.ts
@@ -3,13 +3,13 @@ import { Module } from '@nestjs/common';
 import { ConnectApiKeyService } from '../authn/connect-api-key.service.js';
 import { CloudResolver } from '../connection-status/cloud.resolver.js';
 import { CloudService } from '../connection-status/cloud.service.js';
-import { MothershipHandler } from '../event-handler/mothership.events.js';
+import { MothershipHandler } from './mothership.events.js';
 import { TimeoutCheckerJob } from '../connection-status/timeout-checker.job.js';
-import { MothershipConnectionService } from '../service/connection.service.js';
-import { MothershipGraphqlClientService } from '../service/graphql.client.js';
+import { MothershipConnectionService } from './connection.service.js';
+import { MothershipGraphqlClientService } from './graphql.client.js';
 import { InternalClientService } from '../internal-rpc/internal.client.js';
-import { MothershipSubscriptionHandler } from '../service/mothership-subscription.handler.js';
-import { RemoteAccessModule } from './remote-access.module.js';
+import { MothershipSubscriptionHandler } from './mothership-subscription.handler.js';
+import { RemoteAccessModule } from '../module/remote-access.module.js';
 
 @Module({
     imports: [RemoteAccessModule],
